+++ app/components/search/videoList.js
@@ -2,17 +2,18 @@ import React from 'react';
 
 import VideoListItem from './videoListItem';
 
-const { arrayOf, shape, string } = React.PropTypes;
+const { arrayOf, shape, string, func } = React.PropTypes;
 
-const VideoList = ({ videos }) => {
-    const videoItems = videos.map((video) => {
-        return <VideoListItem key={ video.id } video={ video }/>;
+const VideoList = ({ videos, onSend }) => {
+    const videoItems = videos.map((video, index) => {
+        return <VideoListItem key={ video.id } video={ video } itemType={ index % 2 === 0 ? 'even' : 'odd' }  onSend={ onSend }/>;
     });
 
     return <div className="video-list-container"><ul className="video-list">{ videoItems }</ul></div>;
 };
 
 VideoList.propTypes = {
+    onSend: func,
     videos: arrayOf(
         shape({
             id: string.isRequired,
