--- app/helpers/checkStatus.js
@@ -1,6 +1,0 @@
-export default function (response) {
-    if (response.ok) {
-        return Promise.resolve(response);
-    }
-    return Promise.reject(new Error(response.statusText));
-}
