--- app/helpers/parseJson.js
@@ -1,3 +1,0 @@
-export default function (response) {
-    return response.json();
-}
