+++ app/styles/components/singIn.scss
@@ -38,6 +38,7 @@
   }
 }
 .singing-overlay{
+  z-index: 100;
   background:rgba(0,0,0,0.6);
   position: absolute;
   top: 0;
