+++ app/styles/main.scss
@@ -1,6 +1,7 @@
 $main-color: #0389cb;
 $secondary-color: #e8e8e8;
 $userNameColor: #ffffff;
+$black: black;
 
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:300,400,700');
 @import url('https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css');
